requirements.txt
coremltools==6.2
easydict==1.9
Flask==2.2.2
imutils==0.5.4
ipdb==0.13.11
matplotlib==3.5.2
motmetrics==1.4.0
numpy==1.19.5
onnx==1.12.0
onnxsim==0.4.17
opencv_python==4.6.0.66
pafy==0.5.5
pandas==1.3.5
Pillow==9.4.0
protobuf==4.21.12
PyYAML==6.0
requests==2.28.1
scipy==1.7.3
seaborn==0.11.2
setuptools==67.2.0
thop==0.1.1.post2209072238
torch==1.9.0+cu111
torchfile==0.1.0
torchvision==0.10.0+cu111
tqdm==4.64.0
wandb==0.13.10
